Changes to adding components, added arrays, object composition
Built in components are no longer added using strings, and are instead exposed on the emvy object. This was a dodgy design to begin with, and while this method is more verbose, its also clearer.

Emvy arrays: These provide a way of adding model-like events to normal arrays and normal array operations. Allows a ViewCollection to just use an array

Accompanying emvy arrays is a new method for simple object creation, emvy.create, which can decorate objects with emvy components. This means entirely classless objects and arrays can be used to back Views and ViewCollections.