changed the syntax for stores, so they can push updates
stores functions are now passed with the name to the @store() setter, allowing emvy to call the store with create, update, delete, functions passed through. These can be used within the store to modify models in response to say, socket events, or other pushed, polled, or otherwise changing data.

internally the "rebuilding" flag is set when updating content from the server. This really needs a better name.